slide-sift-backend/requirements.txt
google-generativeai==0.7.2pypdf
flask
flask-cors
gunicorn
python-dotenv
werkzeug>=3.0.0